lambda/requirements.txt
-e .

# aws cdk
aws-cdk.core==1.126.0
aws-cdk.aws_iam==1.126.0
aws-cdk.aws_s3_assets==1.126.0
aws_cdk.aws_apigateway==1.126.0
aws_cdk.aws_lambda==1.126.0
aws_cdk.aws_s3==1.126.0
